* commit 'aac36e919a02f2d0088bc56746b2928c380ce138':
  refactored exception handlers, created entities,changed umico request,response and changed mappers and .etc

+++ src/main/java/az/kapitalbank/marketplace/dto/OrderProductItem.java
@@ -2,6 +2,7 @@ package az.kapitalbank.marketplace.dto;
 
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AccessLevel;
@@ -19,7 +20,7 @@ import lombok.experimental.FieldDefaults;
 public class OrderProductItem {
     @NotNull
     @JsonProperty("product_amount")
-    Integer productAmount;
+    BigDecimal productAmount;
     @NotEmpty
     @JsonProperty("product_id")
     String productId;
